+++ src/plugins/oSnap/components/HandleOutcome/steps/CanProposeToOG.vue
