+++ src/plugins/oSnap/components/HandleOutcome/steps/CanRequestTxExecution.vue
